RubyLexer: Generalized non-expanded delimited string
