+++ joern-cli/frontends/rubysrc2cpg/src/main/scala/io/joern/rubysrc2cpg/parser/NonExpandedDelimitedStringHandling.scala
@@ -1,0 +1,53 @@
+package io.joern.rubysrc2cpg.parser
+
+import scala.collection.mutable
+
+trait NonExpandedDelimitedStringHandling { this: RubyLexerBase =>
+
+  private val delimiters = mutable.Stack[Int]()
+  private var nonExpandedDelimitedStringEndTokenType = 0
+
+  def pushNonExpandedDelimiter(char: Int): Unit = {
+    delimiters.push(char)
+  }
+
+  def popNonExpandedDelimiter(): Unit = {
+    delimiters.pop()
+  }
+
+  def isNonExpandedDelimitersStackEmpty: Boolean = {
+    delimiters.isEmpty
+  }
+
+  def isNonExpandedOpeningDelimiter(char: Int): Boolean = {
+    char == currentOpeningDelimiter()
+  }
+
+  def isNonExpandedClosingDelimiter(char: Int): Boolean = {
+    char == currentClosingDelimiter()
+  }
+  
+  def isNonExpandedDelimiter(char: Int): Boolean = {
+    isNonExpandedOpeningDelimiter(char) || isNonExpandedClosingDelimiter(char)
+  }
+
+  private def currentOpeningDelimiter(): Int = {
+    delimiters.top
+  }
+  
+  def setNonExpandedDelimitedStringEndToken(endTokenType: Int): Unit = {
+    this.nonExpandedDelimitedStringEndTokenType = endTokenType
+  }
+  
+  def getNonExpandedDelimitedStringEndToken: Int = nonExpandedDelimitedStringEndTokenType
+  
+  private def currentClosingDelimiter(): Int =
+    currentOpeningDelimiter() match {
+      case '(' => ')'
+      case '[' => ']'
+      case '{' => '}'
+      case '<' => '>'
+      case c => c
+    }
+
+}
