+++ joern-cli/frontends/rubysrc2cpg/src/main/scala/io/joern/rubysrc2cpg/parser/RubyLexerBase.scala
@@ -9,10 +9,9 @@ abstract class RubyLexerBase(input: CharStream)
     extends Lexer(input)
     with RubyLexerRegexHandling
     with RubyLexerStringInterpolationHandling
-    with RubyLexerQuotedNonExpandedStringHandling
     with RubyLexerQuotedNonExpandedStringArrayHandling
     with RubyLexerQuotedNonExpandedSymbolArrayHandling
-    with RubyLexerQuotedNonExpandedRegularExpressionHandling {
+    with NonExpandedDelimitedStringHandling {
 
   /** The previously (non-WS) emitted token (in DEFAULT_CHANNEL.) */
   protected var previousNonWsToken: Option[Token] = None
